refactor image upload logic to emit
