+++ src/receipts/upload-image/upload-image.component.ts
@@ -1,14 +1,13 @@
-import { tap } from 'rxjs';
-import { FormMode } from 'src/enums/form-mode.enum';
-import { Component, Input, ViewChild } from '@angular/core';
-import { ActivatedRoute } from '@angular/router';
 import {
-  FileData,
-  ReceiptImageService,
-  SnackbarService,
-} from '@receipt-wrangler/receipt-wrangler-core';
-
-import { formatImageData } from '../utils/form.utils';
+  Component,
+  EventEmitter,
+  Input,
+  Output,
+  ViewChild,
+} from '@angular/core';
+import { ActivatedRoute } from '@angular/router';
+import { FileData } from '@receipt-wrangler/receipt-wrangler-core';
+import { FormMode } from 'src/enums/form-mode.enum';
 
 @Component({
   selector: 'app-upload-image',
@@ -22,17 +21,15 @@ export class UploadImageComponent {
 
   @Input() public mode: FormMode = FormMode.view;
 
+  @Output() public fileLoaded: EventEmitter<FileData> = new EventEmitter();
+
   @ViewChild('uploadInput') uploadInput!: any;
 
   public formMode = FormMode;
 
   public acceptFileType: string = 'image/*';
 
-  constructor(
-    private activatedRoute: ActivatedRoute,
-    private receiptImageService: ReceiptImageService,
-    private snackbarService: SnackbarService
-  ) {
+  constructor(private activatedRoute: ActivatedRoute) {
     this.mode = this.activatedRoute.snapshot.data['mode'];
   }
 
@@ -60,38 +57,10 @@ export class UploadImageComponent {
           receiptId: this.receiptId,
         } as any as FileData;
 
-        this.handleFile(fileData);
+        this.fileLoaded.emit(fileData);
       };
 
       reader.readAsBinaryString(f);
-    }
-  }
-
-  private handleFile(fileData: FileData): void {
-    switch (this.mode) {
-      case FormMode.add:
-        this.images.push(fileData);
-        break;
-      case FormMode.edit:
-        const uploadData = formatImageData(
-          fileData,
-          Number.parseInt(this.receiptId ?? '')
-        );
-        this.receiptImageService
-          .uploadReceiptImage(uploadData)
-          .pipe(
-            tap((data: FileData) => {
-              this.snackbarService.success('Successfully uploaded image(s)');
-              fileData.id = data.id;
-              this.images.push(fileData);
-            })
-          )
-          .subscribe();
-        // we can upload each, then
-        break;
-
-      default:
-        break;
     }
   }
 }
